ci: add operational dev and prod environment with docker
The dev and prod environment can now be started within Docker using the `dev.sh` and `prod.sh` scripts.

The dev environment launches the api in watch mode and the next.js website in development mode.
The prod environment launches the api and the next.js website in production mode.

--amend

ci: add operational dev and prod environment with docker

The dev and prod environment can be started within Docker using the `dev.sh` and `prod.sh` scripts.

The dev environment launches the api in watch mode and the next.js website in development mode.
The prod environment launches the api and the next.js website in production mode.